Update TxRequestQueue.java
Set efacloud upload sync timeout to 15 minutes (after discussion with Martin Glade).

This leads to less entries in efaloud.log that the sync has timed out and will be cancelled (but obivously, it is not cancelled due to the bug).

The timeout handling does currently not work, but an update would be rather complex. So this needs to be done in another feature. An issue has been documented on GitHub.